requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
jinja2==3.1.6
python-multipart==0.0.20
reportlab==4.3.1
camel-ai[web_tools]==0.2.44
openai==1.70.0
requests==2.32.3
python-dotenv==1.1.0
